+++ Models/PeeringTask-User.cs
@@ -21,7 +21,7 @@ namespace patools.Models
         [Required]
         public PeeringTaskStates State { get; set; }
 
-        public int? FinalGrade { get; set; }
+        public float? FinalGrade { get; set; }
         public float? SubmissionGrade { get; set; }
         public float? ReviewGrade { get; set; }
